services/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.1",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "orjson>=3.11.2",
    "psycopg>=3.2.9",
    "psycopg-pool>=3.2.6",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "respx>=0.22.0",
    "ruff>=0.12.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "S"]
ignore = ["E501"]

[tool.basedpyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownParameterType = "none"
reportUnknownArgumentType = "none"
reportMissingParameterType = "none"
